Do bit shifting on unsigned types for the sake of -fsanitize=undefined
src/ChangeLog addition:

2024-09-23  Aidan Kehoe  <kehoea@parhasard.net>

	* data.c (Fash):
	Do bit shifting on EMACS_UINTs for the sake of -fsanitize=undefined.
	* regex.c:
	* regex.c (extract_number):
	Do bit shifting on an unsigned type for the sake of
	-fsanitize=undefined.
	* sysdep.c (get_random):
	Ditto.

+++ src/regex.c
@@ -683,7 +683,7 @@ static inline int
 extract_number (re_char *source)
 {
   int leading_byte = SIGN_EXTEND_CHAR (source[1]);
-  return (leading_byte << 8) + (int) ((unsigned char) (source[0]));
+  return ((unsigned) leading_byte << 8) + (int) ((unsigned char) (source[0]));
 }
 
 /* Similar to EXTRACT_NUMBER, but treat the two bytes at SOURCE as an unsigned
